+++ runtime/include/chpl-mem.h
@@ -27,12 +27,9 @@
 #include <string.h>
 #include <assert.h>
 #include "arg.h"
-#include "chpl-comm.h"
 #include "chpl-mem-desc.h"
 #include "chpl-mem-hook.h"
-#include "chpl-topo.h"
 #include "chpltypes.h"
-#include "chpl-tasks.h"
 #include "error.h"
 
 
@@ -123,117 +120,20 @@ void* chpl_mem_realloc(void* memAlloc, size_t size,
 }
 
 static inline
+void* chpl_mem_memalign(size_t boundary, size_t size,
+                        chpl_mem_descInt_t description,
+                        int32_t lineno, int32_t filename) {
+  void* memAlloc;
+  chpl_memhook_malloc_pre(1, size, description, lineno, filename);
+  memAlloc = chpl_memalign(boundary, size);
+  chpl_memhook_malloc_post(memAlloc, 1, size, description, lineno, filename);
+  return memAlloc;
+}
+
+static inline
 void chpl_mem_free(void* memAlloc, int32_t lineno, int32_t filename) {
   chpl_memhook_free_pre(memAlloc, lineno, filename);
   chpl_free(memAlloc);
-}
-
-static inline
-chpl_bool chpl_mem_size_justifies_comm_alloc(size_t size) {
-  //
-  // Don't try to use comm layer allocation unless the size is beyond
-  // the comm layer threshold and enough pages to make localization
-  // worthwhile.
-  //
-  return (size >= chpl_comm_regMemAllocThreshold()
-          && size >= 2 * chpl_comm_regMemHeapPageSize());
-}
-
-static inline
-void* chpl_mem_array_alloc(size_t nmemb, size_t eltSize, c_sublocid_t subloc,
-                           chpl_bool* callAgain, void* repeat_p,
-                           int32_t lineno, int32_t filename) {
-  const size_t size = nmemb * eltSize;
-  void* p;
-
-  //
-  // Temporarily, to support dynamic array registration, we accept a
-  // couple of additional arguments.  On a first call, to allocate,
-  // pass NULL for repeat_p.  This function will return a pointer to
-  // the allocated memory and either true or false in *callAgain.  If
-  // false then the allocation procedure is complete.  If true, then
-  // after initializing the memory the caller should call us again,
-  // repeating the original arguments and passing the pointer returned
-  // by the first call in repeat_p.  We will then call the comm layer
-  // post-alloc function.
-  //
-  if (repeat_p == NULL) {
-    //
-    // Allocate and maybe localize.
-    //
-    chpl_bool do_localize;
-
-    p = NULL;
-    *callAgain = false;
-    if (chpl_mem_size_justifies_comm_alloc(size)) {
-      p = chpl_comm_regMemAlloc(size, CHPL_RT_MD_ARRAY_ELEMENTS,
-                                lineno, filename);
-      if (p != NULL) {
-        *callAgain = true;
-        do_localize = (subloc == c_sublocid_all) ? true : false;
-      }
-    }
-
-    if (p == NULL) {
-      p = chpl_mem_allocMany(nmemb, eltSize, CHPL_RT_MD_ARRAY_ELEMENTS,
-                             lineno, filename);
-      do_localize = (subloc == c_sublocid_all) ? true : false;
-    }
-
-    if (do_localize) {
-      if (isActualSublocID(subloc)) {
-        chpl_topo_setMemLocality(p, size, true, subloc);
-      }
-    }
-  } else {
-    //
-    // do comm layer post-allocation, if we got the memory from there.
-    //
-    p = repeat_p;
-    chpl_comm_regMemPostAlloc(p, size);
-  }
-
-  return p;
-}
-
-static inline
-void* chpl_mem_wide_array_alloc(int32_t dstNode, size_t nmemb, size_t eltSize,
-                                c_sublocid_t subloc,
-                                chpl_bool* callAgain, void* repeat_p,
-                                int32_t lineno, int32_t filename) {
-  if (dstNode != chpl_nodeID)
-    chpl_error("array vector data is not local", lineno, filename);
-  return chpl_mem_array_alloc(nmemb, eltSize, subloc, callAgain, repeat_p,
-                              lineno, filename);
-}
-
-static inline
-void chpl_mem_array_free(void* p,
-                         size_t nmemb, size_t eltSize,
-                         int32_t lineno, int32_t filename) {
-  const size_t size = nmemb * eltSize;
-
-  //
-  // If the size indicates we might have gotten this memory from the
-  // comm layer then try to free it there.  If not, or if so but the
-  // comm layer says it didn't come from there, free it in the memory
-  // layer.
-  //
-  if (chpl_mem_size_justifies_comm_alloc(size)
-      && chpl_comm_regMemFree(p, size)) {
-    return;
-  }
-
-  chpl_mem_free(p, lineno, filename);
-}
-
-static inline
-void chpl_mem_wide_array_free(int32_t dstNode, void* p,
-                              size_t nmemb, size_t eltSize,
-                              int32_t lineno, int32_t filename) {
-  if (dstNode != chpl_nodeID)
-    chpl_error("array vector data is not local", lineno, filename);
-  chpl_mem_array_free(p, nmemb, eltSize, lineno, filename);
 }
 
 // Provide a handle to instrument Chapel calls to memcpy.
